Update 1.8.Tapete.cpp
agora funciona em sistemas operacionais além do ubuntu

+++ 1.8.Tapete.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    unsigned long int l, n;
+    long long int l, n;
     cin >> l >> n;
     cout << n-1+(l-n+1)*(l-n+1);
     return 0;
